libqfqmlwidgets: compiler warning fixed

+++ libqf/libqfqmlwidgets/src/reports/processor/reportpainter.h
@@ -68,7 +68,7 @@ public:
 	Qt::Alignment alignment() const {return (Qt::Alignment)f_layoutSettings.value(LayoutSetting::Alignment, (int)(Qt::AlignLeft | Qt::AlignTop)).toInt();}
 	void setAlignment(ReportItemFrame::HAlignment hal, ReportItemFrame::VAlignment val) {
 		int al = hal | val;
-		if(alignment() != al)
+		if((int)alignment() != al)
 			f_layoutSettings[LayoutSetting::Alignment] = al;
 	}
 	bool isSuppressPrintOut() {return f_layoutSettings.value(LayoutSetting::SuppressPrintOut).toBool();}
